Add basic, unstyled UI
Testing utils also set up in this commit.

Some might prefer to split the second test into a couple of more
focused tests, but I prefer covering an A/C in a single test if
I can without hampering readability. Dealer's choice.

Tests were written first and mapped as closely as possible to the A/Cs.
There wasn't much opportunity to refactor the initial implementation,
but I suspect I'll have an opportunity when I get to the auto-complete
A/Cs.