Hide the search results when search term is deleted
Note the "Can't perform a React state update on an unmounted component"
warning when running tests. I'm sure this is something simple I've
missed, but I'll ignore it in the interest of not taking too long about
this. Interestingly it only occurs when running multiple tests - they
all pass in isolation.